change rain int to float

+++ app/src/main/java/im/kowalczyk/forecastweather/data/ResponseClasses.kt
@@ -9,7 +9,7 @@ data class ForecastResult(val city: City, val list: List<Forecast>)
 data class City(val id: Long, val name: String, val coord: Coordinates, val country: String, val population: Int)
 
 data class Forecast(val dt: Long, val temp: Temperature, val pressure: Float, val humidity: Int, val weather: List<Weather>,
-                    val speed: Float, val deg: Int, val clouds: Int, val rain: Int)
+                    val speed: Float, val deg: Int, val clouds: Int, val rain: Float)
 
 data class Coordinates(val lon: Float, val lat: Float)
